Changed the numbering of the Cells on the Grid
Now in the website the grid numbering starts from 1 and not 0,before it
started from zero.
Many Articles and questions in Discrete Mathematics they tend to start
with 1 and not with zero, so to follow the convention change was made.

Todo: In future we can add a way to dynamically generate the numbers according
to the users need.